refactor: ProblemCatetory.tsx拆分到多个文件
